+++ WForest/src/UI/Interactions/UserInteractionHandler.cs
@@ -1,5 +1,4 @@
 using System.Collections.Generic;
-using System.Linq;
 using Microsoft.Xna.Framework;
 using WForest.Devices;
 using WForest.UI.Props.Interfaces;